+++ src/main/java/umc6/tom/board/repository/PinCommentRepositoryBoard.java
@@ -1,8 +1,8 @@
 package umc6.tom.board.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import umc6.tom.pin.model.Pin;
-import umc6.tom.pin.model.PinComment;
+import umc6.tom.comment.model.Pin;
+import umc6.tom.comment.model.PinComment;
 
 import java.util.List;
