+++ src/main/java/umc6/tom/comment/dto/PinCommentLikeDto.java
@@ -1,0 +1,14 @@
+package umc6.tom.comment.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class PinCommentLikeDto {
+    private Long id;
+    private Long userId;
+    private Long pinCommentId;
+}
