+++ src/main/java/umc6/tom/comment/dto/PinLikeDto.java
@@ -1,0 +1,14 @@
+package umc6.tom.comment.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class PinLikeDto {
+    private Long id;
+    private Long userId;
+    private Long pinId;
+}
