+++ src/main/java/umc6/tom/comment/dto/PinPictureDto.java
@@ -1,0 +1,16 @@
+package umc6.tom.comment.dto;
+
+import lombok.*;
+
+import java.util.List;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class PinPictureDto {
+    private Long id;
+    private Long pinId;
+    private List<String> pic;
+}
