+++ src/main/java/umc6/tom/comment/model/PinCommentLike.java
@@ -1,0 +1,25 @@
+package umc6.tom.comment.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+import umc6.tom.user.model.User;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class PinCommentLike {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "pinComment_id")
+    private PinComment pinComment;
+}
