+++ src/main/java/umc6/tom/comment/model/PinCommentPicture.java
@@ -1,0 +1,23 @@
+package umc6.tom.comment.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class PinCommentPicture {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, length = 100)
+    private String pic;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "pinComment_id")
+    private PinComment pinComment;
+}
