+++ src/main/java/umc6/tom/comment/repository/PinRepository.java
@@ -1,0 +1,9 @@
+package umc6.tom.comment.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import umc6.tom.comment.model.Pin;
+
+@Repository
+public interface PinRepository extends JpaRepository<Pin, Long> {
+}
