--- src/main/java/umc6/tom/pin/dto/PinCommentLikeDto.java
@@ -1,14 +1,0 @@
-package umc6.tom.pin.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class PinCommentLikeDto {
-    private Long id;
-    private Long userId;
-    private Long pinCommentId;
-}
