--- src/main/java/umc6/tom/pin/dto/PinCommentPictureDto.java
@@ -1,14 +1,0 @@
-package umc6.tom.pin.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class PinCommentPictureDto {
-    private Long id;
-    private Long pinCommentId;
-    private String pic;
-}
