--- src/main/java/umc6/tom/pin/dto/PinPictureDto.java
@@ -1,16 +1,0 @@
-package umc6.tom.pin.dto;
-
-import lombok.*;
-
-import java.util.List;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class PinPictureDto {
-    private Long id;
-    private Long pinId;
-    private List<String> pic;
-}
