--- src/main/java/umc6/tom/pin/model/PinComment.java
@@ -1,38 +1,0 @@
-package umc6.tom.pin.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-import umc6.tom.common.BaseEntity;
-import umc6.tom.user.model.User;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class PinComment extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 50)
-    private String comment;
-
-    @OneToMany(mappedBy = "pinComment", cascade = CascadeType.ALL)
-    private List<PinCommentLike> pinCommentLikeList = new ArrayList<>();
-
-    @OneToMany(mappedBy = "pinComment", cascade = CascadeType.ALL)
-    private List<PinCommentPicture> pinCommentPictureList = new ArrayList<>();
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "pin_id")
-    private Pin pin;
-}
