--- src/main/java/umc6/tom/pin/model/PinCommentPicture.java
@@ -1,23 +1,0 @@
-package umc6.tom.pin.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class PinCommentPicture {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 100)
-    private String pic;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "pinComment_id")
-    private PinComment pinComment;
-}
