--- src/main/java/umc6/tom/pin/model/PinPicture.java
@@ -1,23 +1,0 @@
-package umc6.tom.pin.model;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class PinPicture {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 100)
-    private String pic;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "pin_id")
-    private Pin pin;
-}
