--- src/main/java/umc6/tom/pin/repository/PinLikeRepository.java
@@ -1,11 +1,0 @@
-package umc6.tom.pin.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import umc6.tom.pin.model.PinLike;
-import umc6.tom.user.model.User;
-
-@Repository
-public interface PinLikeRepository extends JpaRepository<PinLike, Long> {
-    PinLike findByUser(User user);
-}
